+++ include/rpc/rpc_client.h
@@ -17,6 +17,7 @@
 #include "rpc/rpc_types.h"
 #include "platform/buffer_pool.h"
 #include "transport/endpoint.h"
+#include "transport/message_rejection.h"
 
 #ifdef SOMEIP_STATIC_ALLOC
 #include "static_config.h"
@@ -154,6 +155,17 @@ public:
      * @return true if ready for RPC calls
      */
     bool is_ready() const;
+
+    /**
+     * @brief Observe structural rejections of incoming messages.
+     *
+     * Additive; default is no handler. Invoked on the transport receive
+     * thread. Does not send a SOME/IP error response.
+     *
+     * @implements REQ_TRANSPORT_026
+     */
+    void set_message_rejection_handler(
+        platform::Function<void(const transport::MessageRejectionInfo&)> handler);
 
     /**
      * @brief Get client statistics